# frontend/openapi.json
{
  "components": {
    "schemas": {
...
      "Category": {
        "enum": [
          "BLOQUEO_TARJETA",
          "REPORTE_FRAUDE",
          "CONSULTA_GENERAL",
          "SOLICITUD_CREDITO",
          "BANCA_DIGITAL"
        ],
        "title": "Category",
        "type": "string"
      },
...
      "ExecutiveStatus": {
        "enum": [
          "DISPONIBLE",
          "OCUPADO",
          "INACTIVO"
        ],
        "title": "ExecutiveStatus",
        "type": "string"
      },
      "ExecutiveStatusResult": {
        "properties": {
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/ExecutiveStatus"
          },
          "unassigned_tickets": {
            "default": 0,
            "title": "Unassigned Tickets",
            "type": "integer"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "status",
          "version"
        ],
        "title": "ExecutiveStatusResult",
        "type": "object"
      },
      "ExecutiveStatusUpdate": {
        "properties": {
          "expected_version": {
            "minimum": 1.0,
            "title": "Expected Version",
            "type": "integer"
          },
          "status": {
            "$ref": "#/components/schemas/ExecutiveStatus"
          }
        },
        "required": [
          "status",
          "expected_version"
        ],
        "title": "ExecutiveStatusUpdate",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ManagementCasesResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ManagerialCase"
            },
            "title": "Items",
            "type": "array"
          },
          "page": {
            "title": "Page",
            "type": "integer"
          },
          "page_size": {
            "title": "Page Size",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "page",
          "page_size",
          "total"
        ],
        "title": "ManagementCasesResponse",
        "type": "object"
      },
...
      "ManagementTicketMutation": {
        "properties": {
          "executive_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Executive Id"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "number": {
            "title": "Number",
            "type": "string"
          },
          "priority": {
            "$ref": "#/components/schemas/Priority"
          },
          "status": {
            "$ref": "#/components/schemas/TicketStatus"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "number",
          "executive_id",
          "priority",
          "status",
          "version"
        ],
        "title": "ManagementTicketMutation",
        "type": "object"
      },
...
      "Priority": {
        "enum": [
          "CRITICO",
          "ALTO",
          "MEDIO",
          "BAJO"
        ],
        "title": "Priority",
        "type": "string"
      },
...
      "TicketAssignmentUpdate": {
        "properties": {
          "executive_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Executive Id"
          },
          "expected_version": {
            "minimum": 1.0,
            "title": "Expected Version",
            "type": "integer"
          },
          "reason": {
            "maxLength": 500,
            "minLength": 10,
            "title": "Reason",
            "type": "string"
          }
        },
        "required": [
          "executive_id",
          "expected_version",
          "reason"
        ],
        "title": "TicketAssignmentUpdate",
        "type": "object"
      },
...
      "TicketPriorityUpdate": {
        "properties": {
          "expected_version": {
            "minimum": 1.0,
            "title": "Expected Version",
            "type": "integer"
          },
          "priority": {
            "$ref": "#/components/schemas/Priority"
          },
          "reason": {
            "maxLength": 500,
            "minLength": 10,
            "title": "Reason",
            "type": "string"
          }
        },
        "required": [
          "priority",
          "expected_version",
          "reason"
        ],
        "title": "TicketPriorityUpdate",
        "type": "object"
      },
...
      "TicketStatus": {
        "enum": [
          "PENDIENTE",
          "EN_ATENCION",
          "CERRADO"
        ],
        "title": "TicketStatus",
        "type": "string"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "API del sistema de orquestacion de atencion bancaria",
    "title": "Sistema de Orquestacion Bancaria",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/management/cases": {
      "get": {
        "operationId": "cases_api_v1_management_cases_get",
        "parameters": [
          {
            "in": "query",
            "name": "date_from",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date From"
            }
          },
          {
            "in": "query",
            "name": "date_to",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date To"
            }
          },
          {
            "in": "query",
            "name": "category",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/Category"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Category"
            }
          },
          {
            "in": "query",
            "name": "priority",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/Priority"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Priority"
            }
          },
          {
            "in": "query",
            "name": "executive_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Executive Id"
            }
          },
          {
            "in": "query",
            "name": "unassigned",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Unassigned"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/TicketStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 120,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Page Size",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ManagementCasesResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Cases",
        "tags": [
          "Gestion gerencial"
        ]
      }
    },
    "/api/v1/management/executives/{executive_id}/status": {
      "patch": {
        "operationId": "update_executive_status_api_v1_management_executives__executive_id__status_patch",
        "parameters": [
          {
            "in": "path",
            "name": "executive_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Executive Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExecutiveStatusUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExecutiveStatusResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update Executive Status",
        "tags": [
          "Gestion gerencial"
        ]
      }
    },
...
    "/api/v1/management/tickets/{ticket_id}/assignment": {
      "patch": {
        "operationId": "update_ticket_assignment_api_v1_management_tickets__ticket_id__assignment_patch",
        "parameters": [
          {
            "in": "path",
            "name": "ticket_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Ticket Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TicketAssignmentUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ManagementTicketMutation"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update Ticket Assignment",
        "tags": [
          "Gestion gerencial"
        ]
      }
    },
    "/api/v1/management/tickets/{ticket_id}/priority": {
      "patch": {
        "operationId": "raise_ticket_priority_api_v1_management_tickets__ticket_id__priority_patch",
        "parameters": [
          {
            "in": "path",
            "name": "ticket_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Ticket Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TicketPriorityUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ManagementTicketMutation"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Raise Ticket Priority",
        "tags": [
          "Gestion gerencial"
        ]
      }
    },
...
  }
}